renamed types.py to dimension_types.py
This is to avoid the confusion between dimension_types and data_types
