+++ breadbox/breadbox/compute/dataset_tasks.py
@@ -29,7 +29,7 @@ from breadbox.schemas.dataset import (
 )
 from ..config import get_settings
 from ..service import dataset as dataset_service
-from ..crud import types as type_crud
+from ..crud import dimension_types as type_crud
 from ..crud import group as group_crud
 from ..crud import data_type as data_type_crud
 from ..io.data_validation import validate_and_upload_dataset_files
