+++ breadbox/db_load.py
@@ -9,7 +9,7 @@ import pandas as pd
 from breadbox.service.dataset import add_dimension_type
 
 from breadbox.crud.data_type import get_data_types
-from breadbox.api.types import Settings
+from breadbox.api.dimension_types import Settings
 from breadbox.schemas.types import IdMapping, AnnotationTypeMap
 import os
 from typing import Protocol, Any
