+++ breadbox/tests/api/test_datasets.py
@@ -4,7 +4,7 @@ import uuid
 import numpy as np
 import pandas as pd
 
-from breadbox.crud.types import get_dimension_type
+from breadbox.crud.dimension_types import get_dimension_type
 from ..utils import assert_status_not_ok, assert_status_ok, assert_task_failure
 
 from sqlalchemy import and_
